Update main.py
added "saved_list" and more user customization/interactions. I was going to make saved-list as a text file but can't figure out why python only adds to the file when it is finished executing instead of during: so I made a list for now and commented out the text file codes hoping to return later.